+++ hail/src/main/resources/include/hail/Region2.h
@@ -5,6 +5,8 @@
 #include <vector>
 #include <utility>
 #include "hail/NativeStatus.h"
+#include "hail/NativeObj.h"
+#include "hail/NativePtr.h"
 
 namespace hail {
 
@@ -108,6 +110,30 @@ class RegionPool {
 
 using Region2 = RegionPool::Region;
 using RegionPtr = Region2::SharedPtr;
+
+class ScalaRegionPool : public NativeObj {
+  RegionPool pool_{};
+  public:
+    class Region : public NativeObj {
+      private:
+        RegionPtr region_;
+
+      public:
+        Region(ScalaRegionPool * pool) : region_(pool->pool_.get_region()) { }
+        void clear();
+        inline void align(size_t alignment) { region_->align(alignment); }
+        inline char * allocate(size_t alignment, size_t n) { return region_->allocate(alignment, n); }
+        inline char * allocate(size_t n) { return region_->allocate(n); }
+
+        virtual const char* get_class_name() { return "Region"; }
+        virtual ~Region() { region_ = nullptr; }
+    };
+
+    inline std::shared_ptr<Region> get_region() {
+      return std::make_shared<Region>(this);
+    }
+    virtual const char* get_class_name() { return "RegionPool"; }
+};
 
 }
